+++ lib/fiber_units.rb
@@ -8,6 +8,7 @@ require_relative "fiber_units/conversions/weight_conversion"
 
 require_relative "fiber_units/dimensions/length"
 require_relative "fiber_units/dimensions/weight"
+require_relative "fiber_units/dimensions/count"
 require_relative "fiber_units/dimensions/stitch_count"
 require_relative "fiber_units/dimensions/row_count"