+++ lib/fiber_units/dimensions/row_count.rb
@@ -1,17 +1,4 @@
 module FiberUnits
-  class RowCount
-    attr_reader :value
-
-    def initialize(value, _unit = nil)
-      @value = value
-    end
-
-    def +(other)
-      RowCount.new(value + other.value)
-    end
-
-    def ==(other)
-      other.is_a?(RowCount) && value == other.value
-    end
+  class RowCount < Count
   end
 end